+++ lib/ex_ratatui/native.ex
@@ -14,9 +14,11 @@ defmodule ExRatatui.Native do
       aarch64-unknown-linux-gnu
       aarch64-unknown-linux-musl
       arm-unknown-linux-gnueabihf
+      riscv64gc-unknown-linux-gnu
       x86_64-apple-darwin
       x86_64-pc-windows-gnu
       x86_64-pc-windows-msvc
+      x86_64-unknown-freebsd
       x86_64-unknown-linux-gnu
       x86_64-unknown-linux-musl
     ),